+++ include/render/primitives/Curve.h
@@ -76,6 +76,8 @@ namespace render {
                                        render::State& state) const override;
     PrimitivePacketHit4 intersectPacketHits(const Ray4& rays,
                                             const PrimitivePacketState4& states) const override;
+    PrimitivePacketHit8 intersectPacketHits(const Ray8& rays,
+                                            const PrimitivePacketState8& states) const override;
 
     void forEachCurveOverlaySegment(const CurveOverlaySegmentVisitor& visitor) const override;